Removed wrongly named @profile-pic-tiny-width and moved the users profile details into ProfileCard for easier reusability
