+++ src/components/global.less
@@ -10,12 +10,4 @@
   -moz-osx-font-smoothing: grayscale;
 }
 
-@profile-pic-tiny-width: 80px;
-
-:global .profile-tiny {
-  border-radius: @profile-pic-tiny-width / 2;
-  width: @profile-pic-tiny-width;
-  height: @profile-pic-tiny-width;
-}
-
