+++ src/components/mixins.less
@@ -29,3 +29,7 @@
 
 // Breakpoints 
 @bp-mobile: 800px;
+
+
+// Metrics
+@base-padding: 3rem;
